Enable possibility to set RSU position in GeoNet
This commit enables the possibility to set the RSU/reference position
for GeoNetworking, which was previously fixed to an hardcoded point, thus
not enabling services different than the sample ones.
Now this should be fixed.
This commit also removed two unused GeoNet files.

+++ src/automotive/examples/emu-v2x.cc
@@ -62,7 +62,7 @@ main (int argc, char *argv[])
 
   double emuTime = 100;
 
-  uint16_t numberOfNodes = 0;
+  int numberOfNodes = 0;
   uint32_t nodeCounter = 0;
 
   CommandLine cmd;
