+++ src/automotive/examples/v2v-80211p.cc
@@ -52,7 +52,7 @@ main (int argc, char *argv[])
 
   double simTime = 100;
 
-  uint16_t numberOfNodes;
+  int numberOfNodes;
   uint32_t nodeCounter = 0;
 
   CommandLine cmd;
